Remove circuit breaker

+++ internal/engine/rampup.go
@@ -108,7 +108,7 @@ func (rum *RampUpManager) setAllPublishersToFullRate(publishers []PublisherInter
 func (rum *RampUpManager) updatePublisherRates(publishers []PublisherInterface, progress float64) {
 	for _, pub := range publishers {
 		targetRate := pub.GetTargetRate()
-		currentRate := max(1, min(targetRate, int64(progress*float64(targetRate))))
+		currentRate := max(int64(1), min(targetRate, int64(progress*float64(targetRate))))
 		pub.SetRate(currentRate)
 	}
 }
